Fix double PutObject and presence validation
In a recent change, it was made possible to have two form fields to post a prov or chart. This change in troduced a double post and a double verification when the ChartPostFieldName is the same as the default one.

We fix this by using the filename to index into a map of chartOrProvenaceFile. This allows detecting only one contribution to each chart and prov files.