Register aliases for config keys that don't match the advertised schema
This tries to clean up the situation an bit so that one can write
working configuration files without looking into the code.

I've added a disabled test to check for "invalid" config keys but that
would need renaming of the keys leading to incompatibility with current
config files on user installations.

It would be nice to have a migration path, though.

Fixes: #308
Signed-off-by: jayme-github <tuxnet@gmail.com>