docker: unify Dockerfile and build chartmuseum in Docker context
In order to be able to build a multi-architecture image easily, all
architectures should share the same Dockerfile. For this to be possible
in a portable way, the `chartmuseum` executable should be built inside a
container image.

This commit adds a first step to the container image building which is
the build of chartmuseum itself. In order to keep the image as small as
possible, this is done inside a "builder" image and the produced
executable is then copied into the final image.

Signed-off-by: Gabriel Duque <gabriel@zuh0.com>