Correct README description of config file keys
The README says the conversion between flag name / env var name
and the key used for file-based configuration is simple replacement
of _ with . -- this is not true in many, if not most, cases
(not just the STORAGE flag).

This is demonstrated even in the example that immediately follows
the text: authrealm, not auth.realm for AUTH_REALM, bearerauth for
BEARER_AUTH, and so on.

Further confounding the user, there is no indication (even with debug
on) that the config file contains keys that don't map to any flags.

Signed-off-by: Max Timchenko <max.timchenko@tonal.com>